:file_folder: Set new routes for projects and streams
