+++ pages/index.js
@@ -1,7 +1,4 @@
 import Head from "next/head";
-import Link from "next/link";
-import { getDatabase } from "../lib/notion";
-import { timeAgo } from "../utils/timeAgo";
 
 export const databaseId = process.env.NOTION_DATABASE_ID;
 
@@ -15,12 +12,9 @@ import {
   At,
 } from "phosphor-react";
 
-import BadgeCounter from "../components/BadgeCounter";
-import Card from "../components/Card";
-import CardGrid from "../components/CardGrid";
 import IconSocial from "../components/IconSocial";
 
-const Home = ({ streams }) => {
+const Home = () => {
   return (
     <div className="flex flex-col min-h-screen relative overflow-x-hidden gap-16">
       <Head>
@@ -56,63 +50,8 @@ const Home = ({ streams }) => {
           </div>
         </div>
       </header>
-      <section className="flex flex-col gap-4">
-        <header className="flex flex-col gap-2">
-          <h1 className="text-2xl text-stone-100 font-semibold">
-            Featured Projects
-          </h1>
-          <h4 className="text-lg">
-            I enjoy putting my ideas to practice. Here are some projects I'm
-            proud of.
-          </h4>
-        </header>
-        <CardGrid>
-          <Card
-            title="Laboratório Onírico"
-            description="Art generated by code."
-          />
-        </CardGrid>
-      </section>
-      <section className="flex flex-col gap-4">
-        <header className="flex flex-col gap-2">
-          <div className="flex items-start gap-2">
-            <h1 className="text-2xl text-stone-100 font-semibold">Streams</h1>
-            <BadgeCounter number={streams.length} />
-          </div>
-          <h4 className="text-lg">
-            Streams are the conversations between me and my brain. It's like
-            we're text messaging each other.
-          </h4>
-        </header>
-        <CardGrid>
-          {streams
-            .map((stream) => (
-              <Link href={`/streams/${stream.id}`}>
-                <a>
-                  <Card
-                    key={stream.id}
-                    title={stream.properties.Name.title[0].text.content}
-                    description={timeAgo(stream.created_time)}
-                  />
-                </a>
-              </Link>
-            ))
-            .reverse()}
-        </CardGrid>
-      </section>
     </div>
   );
 };
-
-export async function getStaticProps() {
-  const database = await getDatabase(databaseId);
-
-  return {
-    props: {
-      streams: database,
-    },
-    revalidate: 60,
-  };
-}
 
 export default Home;
